+++ postgresql-patroni/postgresql-patroni.d/local/share/cook/bin/configure-patroni-postgres.sh
@@ -19,6 +19,50 @@ mkdir -p /usr/local/etc/patroni/
 # safe(r) separator for sed
 sep=$'\001'
 
+echo "Cloning consul-template rc scripts"
+cp -a /usr/local/etc/rc.d/consul-template \
+  /usr/local/etc/rc.d/consul-template-postgres
+sed -i '' 's/consul_template/consul_template_postgres/g' \
+  /usr/local/etc/rc.d/consul-template-postgres
+sed -i '' 's/consul-template/consul-template-postgres/g' \
+  /usr/local/etc/rc.d/consul-template-postgres
+ln -s /usr/local/bin/consul-postgres \
+  /usr/local/bin/consul-template-postgres
+
+echo "Writing consul-template-postgres config"
+mkdir -p /usr/local/etc/consul-template-postgres.d
+
+TOKEN=$(/bin/cat /mnt/postgrescerts/unwrapped.token)
+
+< "$TEMPLATEPATH/consul-template-postgres.hcl.in" \
+  > /usr/local/etc/consul-template-postgres.d/consul-template-postgres.hcl
+chmod 600 \
+  /usr/local/etc/consul-template-postgres.d/consul-template-postgres.hcl
+echo "s${sep}%%token%%${sep}$TOKEN${sep}" | sed -i '' -f - \
+  /usr/local/etc/consul-template-postgres.d/consul-template-postgres.hcl
+
+< "$TEMPLATEPATH/postgres.crt.tpl.in" \
+  sed "s${sep}%%ip%%${sep}$IP${sep}g" | \
+  sed "s${sep}%%nodename%%${sep}$NODENAME${sep}g" | \
+  sed "s${sep}%%bttl%%${sep}$BTTL${sep}g" \
+  > /mnt/templates/postgres.crt.tpl
+
+< "$TEMPLATEPATH/postgres.key.tpl.in" \
+  sed "s${sep}%%ip%%${sep}$IP${sep}g" | \
+  sed "s${sep}%%nodename%%${sep}$NODENAME${sep}g" | \
+  sed "s${sep}%%bttl%%${sep}$BTTL${sep}g" \
+  > /mnt/templates/postgres.key.tpl
+
+< "$TEMPLATEPATH/postgres-ca.crt.tpl.in" \
+  sed "s${sep}%%ip%%${sep}$IP${sep}g" | \
+  sed "s${sep}%%nodename%%${sep}$NODENAME${sep}g" | \
+  sed "s${sep}%%bttl%%${sep}$BTTL${sep}g" \
+  > /mnt/templates/postgres-ca.crt.tpl
+
+echo "Enabling and starting consul-template-postgres"
+sysrc consul_template_postgres_syslog_output_enable=YES
+service consul-template-postgres enable
+
 # setup patroni.yml by updating variables
 < "$TEMPLATEPATH/patroni.yml.in" \
   sed "s${sep}%%nodename%%${sep}$NODENAME${sep}g" | \