+++ postgresql-patroni/postgresql-patroni.d/local/share/cook/bin/unwrap-postgres-credentials.sh
@@ -16,7 +16,7 @@ if [ ! -s /mnt/postgrescerts/unwrapped.token ]; then
     WRAPPED_TOKEN=$(< /mnt/postgrescerts/credentials.json \
       jq -re .wrapped_token)
     < /mnt/postgrescerts/credentials.json \
-      jq -re .cert >/mnt/postgrescerts/agent.crt
+      jq -re .cert >/mnt/postgrescerts/client.crt
     < /mnt/postgrescerts/credentials.json \
       jq -re .ca >/mnt/postgrescerts/ca.crt
     < /mnt/postgrescerts/credentials.json \
@@ -27,7 +27,7 @@ if [ ! -s /mnt/postgrescerts/unwrapped.token ]; then
       jq -re .ca_root >/mnt/postgrescerts/ca_root.crt
     umask 177
     < /mnt/postgrescerts/credentials.json \
-      jq -re .key >/mnt/postgrescerts/agent.key
+      jq -re .key >/mnt/postgrescerts/client.key
     < /mnt/postgrescerts/credentials.json \
       jq -re .gossip_key >/mnt/postgrescerts/gossip.key
     < /mnt/postgrescerts/credentials.json \