+++ postgresql-patroni/postgresql-patroni.d/local/share/cook/bin/verifynode.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 /usr/local/bin/curl -s \
  --cacert /mnt/postgrescerts/ca_chain.crt \
- --cert /mnt/postgrescerts/agent.crt \
- --key /mnt/postgrescerts/agent.key \
+ --cert /mnt/postgrescerts/client.crt \
+ --key /mnt/postgrescerts/client.key \
  https://127.0.0.1:8008/patroni | /usr/local/bin/jq .